pypic_spacecharge: self-consistent PIC space charge now uses FP64
By default, the PIC atomicAdd variant will use FP64 instead of
FP32 precision, after a bug in PyPIC (introduced during recent
changes) has been fixed.
Timings of a PIC space charge track (1e6 particles) for the P100:
- FP32: 3.78ms
- FP64: 3.79ms
- sorted approach: 8.05ms
Accuracy for adding all kicks (i.e. the influence on the centroid
which ideally due to Newton's 3rd law should be identically zero):
- FP32: <5e-07
- FP64: <5e-14
- sorted approach: 5<e-14
The kicks had an order of <2e-4.

NB: the FP32 atomicAdd version can produce mesh_charges, whose
sum shows an error on the order of 1e-7 (i.e. the first decimal
after the comma is off if you use 1e6 macro-particles).
This is to be expected within floating point precision.
FP64 is much more accurate showing an error well outside the
usually employed range of number of macro-particles.

NB2: FP64 atomicAdd is hardware accelerated on the P100.
FP32 atomicAdd should be twice as quick than FP64, however
our algorithm casts FP64 arrays to FP32 which takes of
the same order of magnitude time as the atomicAdd kernel itself.
Therefore, on the P100 both FP64 and FP32 are approximately
equally quick.
On older architectures, FP64 atomicAdd is software emulated
and therefore much slower, so here you will pay a price
for the increased accuracy.

NB3: for space charge, the 1e-14 centroid kick can add up
over consecutive kicks. Consider Brownian motion of the
centroid (i.e. the error scales with the square root of steps).
Then, O(1e3) kicks per turn for O(1e5) turns yield O(1e8)
kicks for which the error will therefore be O(1e4) times
O(1e-14), so the absolute error after 100'000 turns on
the centroid kick will be O(1e-10). If x' and y' are of the
order of O(1e-6) then the influence of precision error
on the beam centroid is negligible for double precision.
However, single precision FP32 can have a huge impact on
the centroid, so avoid atomicAdd FP32 for space charge
simulations where the centroid motion is relevant (e.g.
for instabilities inflicted by machine impedances).

==> use FP64 atomicAdd particles_to_mesh for space charge
on new GPU architectures.

+++ PyHEADTAIL/spacecharge/pypic_spacecharge.py
@@ -47,7 +47,7 @@ class SpaceChargePIC(Element):
     '''
 
     def __init__(self, length, pypic_algorithm, sort_particles=False,
-                 pic_dtype=np.float32, *args, **kwargs):
+                 pic_dtype=np.float64, *args, **kwargs):
         '''Arguments:
             - length: interaction length over which the space charge
               force is integrated.
